refactor code et fix tele autoplay
