+++ app/Models/Tv.php
